backend/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
azure-ai-documentintelligence==1.0.0b1
azure-core==1.30.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.109.1
h11==0.14.0
idna==3.6
isodate==0.6.1
pydantic==2.6.0
pydantic_core==2.16.1
python-dotenv==1.0.1
python-multipart==0.0.7
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.35.1
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.27.0.post1
